Allow extra query parameters to be specified
The `Query` class now provides untyped accessors to a dictionary of extra parameters. Those parameters are appended to the query string after all the typed parameters. This allows using query parameters not supported by the current version of the API client.

The subscript operator is provided as an alias to the accessor methods.

Fixes #25.